add tests, doc, refactor
