+++ core/api/peoplefinder/team.py
@@ -1,6 +1,7 @@
 from ninja import Router
 
 from core import services as core_services
+from core.schemas import Error
 from core.schemas.peoplefinder.team import PeopleFinderTeamSchema
 
 
@@ -13,7 +14,13 @@ router.add_router("team", team_router)
     "all",
     response={
         200: PeopleFinderTeamSchema,
+        500: Error,
     },
 )
-def get_all_teams(request):
-    return 200, core_services.get_peoplefinder_team_hierarchy()
+def get_hierarcy_of_all_teams(request):
+    try:
+        return 200, core_services.get_peoplefinder_team_hierarchy()
+    except Exception as unknown_error:
+        return 500, {
+            "message": f"Could not get the team hierarchy, reason: {unknown_error}"
+        }
