+++ core/schemas/peoplefinder/team.py
@@ -1,5 +1,3 @@
-from typing import List
-
 from ninja import Field, Schema
 
 
@@ -7,4 +5,4 @@ class PeopleFinderTeamSchema(Schema):
     slug: str = Field(alias="slug")
     name: str = Field(alias="name")
     abbreviation: str = Field(alias="abbreviation")
-    children: List["PeopleFinderTeamSchema"]
+    children: list["PeopleFinderTeamSchema"]
