+++ docs/apis/people-finder.md
@@ -12,6 +12,8 @@ The ID service exposes an API that provides read and edit functionality designed
 `/person/<slug>` : updates an existing people finder profile and returns a full people finder profile as response
 #### Create Profile:
 `/person` : creates a new people finder profile and returns a full people finder profile as response
+#### Get Teams:
+`/team/all` : returns all teams in the team hierarchy
 
 
 #### Reference data can be accessed via `/api/peoplefinder/reference/<endpoint>`. Here is the list of reference endpoints:
@@ -23,5 +25,6 @@ The ID service exposes an API that provides read and edit functionality designed
 - `learning_interests` : list of learning interests
 - `professions` : list of professions
 - `grades` : list of grades
-- `key_skills` : list of key skills- `additional_roles` list of additional roles
+- `key_skills` : list of key skills
+- `additional_roles` list of additional roles
